+++ R/dataframe.R
@@ -1,8 +1,10 @@
 #' @rdname lime
+#' @name lime
 #' @param bin_continuous Should continuous variables be binned when making the explanation
 #' @param n_bins The number of bins for continuous variables if `bin_continuous = TRUE`
 #' @param quantile_bins Should the bins be based on `n_bins` quantiles or spread evenly over the range of the training data
-#' @importFrom stats predict sd quantile
+#' @param use_density If `bin_continuous = FALSE` should continuous data be sampled using a kernel density estimation. If not, continuous features are expected to follow a normal distribution.
+#' @importFrom stats predict sd quantile density
 #' @export
 #'
 #' @examples
@@ -19,22 +21,27 @@
 #' # This can now be used together with the explain method
 #' explain(iris_test, explanation, n_labels = 1, n_features = 2)
 #'
-lime.data.frame <- function(x, model, bin_continuous = TRUE, n_bins = 4, quantile_bins = TRUE, ...) {
+lime.data.frame <- function(x, model, bin_continuous = TRUE, n_bins = 4, quantile_bins = TRUE, use_density = TRUE, ...) {
   explainer <- c(as.list(environment()), list(...))
   explainer$x <- NULL
   explainer$feature_type <- setNames(sapply(x, function(f) {
     if (is.integer(f)) {
-      'integer'
+      if (length(unique(f)) == 1) 'constant' else 'integer'
     } else if (is.numeric(f)) {
-      'numeric'
+      if (length(unique(f)) == 1) 'constant' else 'numeric'
     } else if (is.character(f)) {
       'character'
     } else if (is.factor(f)) {
       'factor'
+    } else if (inherits(f, 'Date') || inherits(f, 'POSIXt')) {
+      'date_time'
     } else {
       stop('Unknown feature type', call. = FALSE)
     }
   }), names(x))
+  if (any(explainer$feature_type == 'constant')) {
+    warning('Data contains numeric columns with zero variance', call. = FALSE)
+  }
   explainer$bin_cuts <- setNames(lapply(seq_along(x), function(i) {
     if (explainer$feature_type[i] %in% c('numeric', 'integer')) {
       if (quantile_bins) {
@@ -52,27 +59,33 @@ lime.data.frame <- function(x, model, bin_continuous = TRUE, n_bins = 4, quantil
       integer = ,
       numeric = if (bin_continuous) {
         table(cut(x[[i]], unique(explainer$bin_cuts[[i]]), labels = FALSE, include.lowest = TRUE))/nrow(x)
+      } else if (use_density) {
+        density(x[[i]])
       } else {
         c(mean = mean(x[[i]], na.rm = TRUE), sd = sd(x[[i]], na.rm = TRUE))
       },
       character = ,
-      factor = table(x[[i]])/nrow(x)
+      factor = table(x[[i]])/nrow(x),
+      NA
     )
   }), names(x))
   structure(explainer, class = c('data_frame_explainer', 'explainer', 'list'))
 }
 #' @rdname explain
+#' @name explain
 #'
 #' @param dist_fun The distance function to use for calculating the distance
-#' from the observation to the permutations. Will be forwarded to
+#' from the observation to the permutations. If `dist_fun = 'gower'` (default) it
+#' will use [gower::gower_dist()]. Otherwise it will be forwarded to
 #' [stats::dist()]
 #' @param kernel_width The width of the exponential kernel that will be used to
-#' convert the distance to a similarity.
+#' convert the distance to a similarity in case `dist_fun != 'gower'`.
 #'
+#' @importFrom gower gower_dist
 #' @export
 explain.data.frame <- function(x, explainer, labels = NULL, n_labels = NULL,
                                n_features, n_permutations = 5000,
-                               feature_select = 'auto', dist_fun = 'euclidean',
+                               feature_select = 'auto', dist_fun = 'gower',
                                kernel_width = NULL, ...) {
   assert_that(is.data_frame_explainer(explainer))
   m_type <- model_type(explainer)
@@ -94,16 +107,22 @@ explain.data.frame <- function(x, explainer, labels = NULL, n_labels = NULL,
   kernel <- exp_kernel(kernel_width)
 
   case_perm <- permute_cases(x, n_permutations, explainer$feature_distribution,
-                             explainer$bin_continuous, explainer$bin_cuts)
+                             explainer$bin_continuous, explainer$bin_cuts,
+                             explainer$use_density)
   case_res <- predict_model(explainer$model, case_perm, type = o_type)
   case_res <- set_labels(case_res, explainer$model)
   case_ind <- split(seq_len(nrow(case_perm)), rep(seq_len(nrow(x)), each = n_permutations))
   res <- lapply(seq_along(case_ind), function(ind) {
     i <- case_ind[[ind]]
+    if (dist_fun == 'gower') {
+      sim <- 1 - gower_dist(case_perm[i[1], , drop = FALSE], case_perm[i, , drop = FALSE])
+    }
     perms <- numerify(case_perm[i, ], explainer$feature_type, explainer$bin_continuous, explainer$bin_cuts)
-    dist <- c(0, dist(feature_scale(perms, explainer$feature_distribution, explainer$feature_type, explainer$bin_continuous),
-                      method = dist_fun)[seq_len(n_permutations-1)])
-    res <- model_permutations(as.matrix(perms), case_res[i, , drop = FALSE], kernel(dist), labels, n_labels, n_features, feature_select)
+    if (dist_fun != 'gower') {
+      sim <- kernel(c(0, dist(feature_scale(perms, explainer$feature_distribution, explainer$feature_type, explainer$bin_continuous),
+                        method = dist_fun)[seq_len(n_permutations-1)]))
+    }
+    res <- model_permutations(as.matrix(perms), case_res[i, , drop = FALSE], sim, labels, n_labels, n_features, feature_select)
     res$feature_value <- unlist(case_perm[i[1], res$feature])
     res$feature_desc <- describe_feature(res$feature, case_perm[i[1], ], explainer$feature_type, explainer$bin_continuous, explainer$bin_cuts)
     guess <- which.max(abs(case_res[i[1], ]))
@@ -129,6 +148,8 @@ numerify <- function(x, type, bin_continuous, bin_cuts) {
   setNames(as.data.frame(lapply(seq_along(x), function(i) {
     if (type[i] %in% c('character', 'factor')) {
       as.numeric(x[[i]] == x[[i]][1])
+    } else if (type[i] == 'date_time' || type[i] == 'constant') {
+      rep(0, nrow(x))
     } else {
       if (bin_continuous) {
         cuts <- bin_cuts[[i]]