+++ R/plot.R
@@ -5,14 +5,16 @@
 #' is shown with its weight, thus giving the importance of the feature in the
 #' label prediction.
 #'
-#' @param explanation A `data.frame` as returned by an explanation function.
+#' @param explanation A `data.frame` as returned by [explain()].
 #'
 #' @param ncol The number of columns in the facetted plot
 #'
-#' @return A ggplot2 object
+#' @return A `ggplot` object
 #'
 #' @import ggplot2
 #' @export
+#'
+#' @family explanation plots
 #'
 #' @examples
 #' # Create some explanations
@@ -52,6 +54,61 @@ plot_features <- function(explanation, ncol = 2) {
     scale_x_discrete(labels = function(lab) substr(lab, desc_width+1, nchar(lab))) +
     labs(y = 'Weight', x = 'Feature', fill = '') +
     theme_lime()
+}
+#' Plot a condensed overview of all explanations
+#'
+#' This function produces a facetted heatmap visualisation of all
+#' case/label/feature combinations. Compared to [plot_features()] it is much
+#' more condensed, thus allowing for an overview of many explanations in one
+#' plot. On the other hand it is less useful for getting exact numerical
+#' statistics of the explanation.
+#'
+#' @param explanation A `data.frame` as returned by [explain()].
+#' @param ... Parameters passed on to [ggplot2::facet_wrap()]
+#'
+#' @return A `ggplot` object
+#'
+#' @import ggplot2
+#' @export
+#'
+#' @family explanation plots
+#'
+#' @examples
+#' # Create some explanations
+#' library(MASS)
+#' iris_test <- iris[1, 1:4]
+#' iris_train <- iris[-1, 1:4]
+#' iris_lab <- iris[[5]][-1]
+#' model <- lda(iris_train, iris_lab)
+#' explanation <- lime(iris_train, model)
+#' explanations <- explain(iris_test, explanation, n_labels = 1, n_features = 2)
+#'
+#' # Get an overview with the standard plot
+#' plot_explanations(explanations)
+plot_explanations <- function(explanation, ...) {
+  num_cases <- unique(suppressWarnings(as.numeric(explanation$case)))
+  if (!anyNA(num_cases)) {
+    explanation$case <- factor(explanation$case, levels = as.character(sort(num_cases)))
+  }
+  explanation$feature_desc <- factor(
+    explanation$feature_desc,
+    levels = rev(unique(explanation$feature_desc[order(explanation$feature, explanation$feature_value)]))
+  )
+  p <- ggplot(explanation, aes_(~case, ~feature_desc)) +
+    geom_tile(aes_(fill = ~feature_weight)) +
+    scale_x_discrete('Case', expand = c(0, 0)) +
+    scale_y_discrete('Feature', expand = c(0, 0)) +
+    scale_fill_gradient2('Feature\nweight', low = '#8e0152', mid = '#f7f7f7', high = '#276419') +
+    theme_lime() +
+    theme(panel.border = element_rect(fill = NA, colour = 'grey60', size = 1),
+          panel.grid = element_blank(),
+          legend.position = 'right',
+          axis.text.x = element_text(angle = 45, hjust = 1, vjust = 1))
+  if (is.null(explanation$label)) {
+    p
+  } else {
+    p + facet_wrap(~label, ...)
+  }
 }
 
 theme_lime <- function(...) {